fix: got client-side working with seperated backend
