+++ client/package.json
@@ -4,17 +4,23 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack --watch"
+    "dev-server": "ttab npx nodemon ./server.js localhost:80",
+    "watch-bundle": "ttab webpack --watch",
+    "start": "webpack --watch",
+    "build": "webpack",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/events": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^4.14.0",
     "@typescript-eslint/parser": "^4.14.0",
     "eslint": "^7.18.0",
     "nodemon": "^2.0.6",
     "npm-run-all": "^4.1.5",
     "ts-loader": "^8.0.4",
+    "ttab": "^0.7.0",
     "typescript": "^4.1.3",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
@@ -23,6 +29,7 @@
     "@types/lodash": "^4.14.168",
     "@types/socket.io-client": "^1.4.34",
     "lodash": "^4.17.20",
+    "socket.io": "^3.1.1",
     "socket.io-client": "^3.1.1",
     "socketio": "^1.0.0"
   }
