+++ client/src/index.ts
@@ -1,6 +1,7 @@
 import { Images } from './images/Images'
 import { Map } from './map/Map'
 import { MessageDistributor } from './managers/MessageDistributor'
+import io from 'socket.io-client'
 
 // console.log('hoi')
 export const images = new Images()
