--- client/test.js
@@ -1,3 +1,0 @@
-const path = require('path')
-
-console.log(path.resolve(__dirname + '/dist/'))
