--- package.json
@@ -1,36 +1,0 @@
-{
-  "name": "boiler-project",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "dev-server": "ttab npx nodemon ./server.js localhost:80",
-    "watch-bundle": "ttab webpack --watch",
-    "start": "run-s dev-server watch-bundle",
-    "build": "webpack",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "@typescript-eslint/parser": "^4.14.0",
-    "eslint": "^7.18.0",
-    "nodemon": "^2.0.6",
-    "npm-run-all": "^4.1.5",
-    "ts-loader": "^8.0.4",
-    "ttab": "^0.7.0",
-    "typescript": "^4.0.3",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
-  },
-  "dependencies": {
-    "@types/lodash": "^4.14.168",
-    "@types/socket.io-client": "^1.4.34",
-    "lodash": "^4.17.20",
-    "lodash.uniqueid": "^4.0.1",
-    "socket.io": "^2.3.0",
-    "socketio": "^1.0.0",
-    "webpack-dev-server": "^3.11.0"
-  }
-}
