+++ server/package.json
@@ -1,23 +1,28 @@
 {
   "name": "server",
   "version": "1.0.0",
+  "type": "module",  
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon src/app.ts",
+    "dev": "nodemon src/server.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/socket.io": "^2.1.13",
     "dotenv": "^8.2.0",
     "express": "^4.17.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.11",
+    "@types/lodash.uniqueid": "^4.0.6",
     "@types/node": "^14.14.25",
+    "lodash.uniqueid": "^4.0.1",
     "nodemon": "^2.0.7",
+    "socket.io": "^2.3.0",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3"
   }
