--- src/index.ts
@@ -1,7 +1,0 @@
-import { Images } from './images/Images'
-import { Map } from './map/Map'
-import { MessageDistributor } from './managers/MessageDistributor'
-
-export const images = new Images()
-export const map = new Map()
-export const messageDistributor = new MessageDistributor()
