--- src/players/Enemy.ts
@@ -1,9 +1,0 @@
-import { ClientInfo } from '../managers/MessageDistributor'
-import { images } from '../index'
-import { Character } from './Character'
-
-export class Enemy extends Character {
-    constructor(protected clientInfo: ClientInfo) {
-        super(clientInfo, images.getImage('player'), 'red')
-    }
-}
