--- src/utils/messages.ts
@@ -1,3 +1,0 @@
-export enum error {
-    noImageFound = "Can't find image"
-}
