--- webpack.config.js
@@ -1,22 +1,0 @@
-const path = require('path')
-
-module.exports = {
-    entry: './src/index.ts',
-    module: {
-        rules: [
-            {
-                test: /\.ts$/,
-                use: 'ts-loader',
-                include: [path.resolve(__dirname, 'src')]
-            }
-        ]
-    },
-    resolve: {
-        extensions: ['.ts', '.js']
-    },
-    output: {
-        publicPath: '/dist/',
-        path: path.resolve(__dirname, './dist'),
-        filename: 'bundle.js',
-    },
-}
